Add files via upload
This program is an example of blocking specific websites during working hours. In order to get this program to work in the background change the file extension to .pyw and set up a task in Windows. Of course you can always change your definition of working hours too. The initial working hours are set to 8am - 5pm.